requirements.txt
azure-cosmos
azure-mgmt-cosmosdb
azure.identity
azure-storage-file-datalake
pandas
pytest
